+++ app/views/events/payments/_payee_picker.html.erb
@@ -18,7 +18,7 @@
     </a>
 
     <% if show_manual %>
-      <%= link_to event_payments_new_path(event_id: event.slug, manual: true), class: "btn btn--list-item border-t rounded-none", data: { turbo_frame: "_top", "payee-picker-target": "searchHidden" } do %>
+      <%= link_to new_event_payment_path(event_id: event.slug, manual: true), class: "btn btn--list-item border-t rounded-none", data: { turbo_frame: "_top", "payee-picker-target": "searchHidden" } do %>
         <div class="w-10 h-10 shrink-0 rounded-full border-2 border-current flex items-center justify-center muted">
           <%= inline_icon "edit", size: 18 %>
         </div>
