+++ app/views/events/payments/new.html.erb
@@ -1,6 +1,6 @@
 <% title "New payment" %>
 <% subtitle "From #{@event.name}" %>
-<% @back_url = event_payments_path(event_id: @event.slug) %>
+<% @back_url = new_event_payment_path(event_id: @event.slug) %>
 
 <% content_for :table_of_contents do %>
   <%= render "events/transfers/step_item", number: 1, label: "Recipient" %>
